Add "Canceled" status support
This commit adds support for builds and deploys to
have a canceled state.

Closes #11
Closes #17
