+++ web/src/styles/custom.scss
@@ -9,6 +9,7 @@ $yellow: #ffc107 !default;
 $green: #66BB6A !default;
 $teal: #20c997 !default;
 $cyan: #29B6F6 !default;
+$secondary: #999999 !default;
 
 @import '~bootstrap';
 @import '~bootstrap-vue';
